feat: enhance payment processing flow with new components and state management
- Added state management for payment processing and detected payment method in the order component.
- Implemented CountdownTimer and PaymentWaiting components to improve user experience during payment processing.
- Updated Step3 component to handle new payment states and integrate the countdown timer for payment expiration.
- Enhanced WalletPaymentSection to conditionally render based on payment method liquidity.
- Improved logging for payment state changes and processing status.
