+++ src/routes/order-new-channel/Step3.tsx
@@ -1,4 +1,3 @@
-import { Info } from 'lucide-react'
 import React, { useState, useCallback, useEffect } from 'react'
 import { ClipLoader } from 'react-spinners'
 import { toast } from 'react-toastify'
@@ -21,154 +20,9 @@ import {
   PaymentStatusDisplay,
   WalletPaymentSection,
   OrderProcessingDisplay,
+  CountdownTimer,
+  PaymentWaiting,
 } from './components'
-
-// Countdown Timer Component
-interface CountdownTimerProps {
-  expiresAt: string
-  onExpiry?: () => void
-}
-
-const CountdownTimer: React.FC<CountdownTimerProps> = ({
-  expiresAt,
-  onExpiry,
-}) => {
-  const expiryDate = new Date(expiresAt)
-  const now = new Date()
-  const initialSecondsRef = React.useRef(
-    Math.max(0, Math.floor((expiryDate.getTime() - now.getTime()) / 1000))
-  )
-  const [countdown, setCountdown] = useState(initialSecondsRef.current)
-
-  useEffect(() => {
-    if (countdown <= 0) {
-      onExpiry?.()
-      return
-    }
-
-    const intervalId = setInterval(() => {
-      setCountdown((prev) => {
-        if (prev <= 1) {
-          clearInterval(intervalId)
-          onExpiry?.()
-          return 0
-        }
-        return prev - 1
-      })
-    }, 1000)
-
-    return () => clearInterval(intervalId)
-  }, [countdown, onExpiry])
-
-  const formatCountdown = (seconds: number) => {
-    const minutes = Math.floor(seconds / 60)
-    const remainingSeconds = seconds % 60
-    return `${minutes.toString().padStart(2, '0')}:${remainingSeconds.toString().padStart(2, '0')}`
-  }
-
-  return (
-    <div className="bg-gray-800/70 p-4 rounded-xl border border-gray-700">
-      <div className="flex items-center justify-between mb-3">
-        <h4 className="text-sm text-gray-400">Payment Expires In</h4>
-        <div className="flex items-center">
-          <div className="w-3 h-3 border-2 border-blue-500 border-t-transparent rounded-full animate-spin mr-2"></div>
-          <span className="text-white font-mono bg-gray-700/70 px-3 py-1 rounded-lg text-lg">
-            {formatCountdown(countdown)}
-          </span>
-        </div>
-      </div>
-
-      <div className="w-full bg-gray-700/50 rounded-full h-2">
-        <div
-          className="h-2 rounded-full transition-all duration-1000 ease-linear"
-          style={{
-            backgroundColor:
-              countdown < 60
-                ? '#EF4444'
-                : countdown < 180
-                  ? '#F59E0B'
-                  : '#3B82F6',
-            width: `${(countdown / initialSecondsRef.current) * 100}%`,
-          }}
-        ></div>
-      </div>
-
-      {countdown < 180 && (
-        <div className="mt-3 flex items-start">
-          <div className="text-yellow-400 mr-2 mt-0.5">
-            <Info size={16} />
-          </div>
-          <p className="text-gray-400 text-xs">
-            {countdown < 60
-              ? 'Payment expires soon!'
-              : 'Payment expiring soon.'}
-          </p>
-        </div>
-      )}
-    </div>
-  )
-}
-
-// PaymentWaiting component
-interface PaymentWaitingProps {
-  paymentURI: string
-  currentPayment: any
-  bitcoinUnit: string
-  paymentMethod: 'lightning' | 'onchain'
-  order: Lsps1CreateOrderResponse | null
-  handleCopy: () => void
-}
-
-const PaymentWaiting: React.FC<PaymentWaitingProps> = ({
-  paymentURI,
-  currentPayment,
-  bitcoinUnit,
-  paymentMethod,
-  order,
-  handleCopy,
-}) => {
-  return (
-    <div className="w-full space-y-6">
-      {/* QR Code Section */}
-      {order && (
-        <QRCodePayment
-          bitcoinUnit={bitcoinUnit}
-          currentPayment={currentPayment}
-          onCopy={handleCopy}
-          order={order}
-          paymentMethod={paymentMethod}
-          paymentURI={paymentURI}
-        />
-      )}
-
-      {/* Payment Status Section */}
-      <div className="space-y-4">
-        <div className="bg-gray-800/70 p-5 rounded-xl border border-gray-700">
-          <div className="flex items-center gap-3 mb-3">
-            <h4 className="text-sm text-gray-400">Payment Status</h4>
-            <div className="flex items-center ml-auto">
-              <div className="w-4 h-4 border-2 border-blue-500 border-t-transparent rounded-full animate-spin mr-2"></div>
-              <span className="text-blue-400 text-sm">Waiting for payment</span>
-            </div>
-          </div>
-        </div>
-
-        <div className="p-4 bg-blue-500/10 border border-blue-500/20 rounded-lg">
-          <div className="flex items-start">
-            <div className="text-blue-400 mr-3 mt-0.5">
-              <Info size={18} />
-            </div>
-            <p className="text-gray-400 text-sm">
-              Keep this window open. The QR code can be scanned to make payment.
-              <br />
-              This page will update automatically once payment is detected.
-            </p>
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
 
 interface StepProps {
   onBack: () => void
@@ -177,6 +31,8 @@ interface StepProps {
   paymentStatus: 'success' | 'error' | 'expired' | null
   orderPayload?: any
   paymentReceived?: boolean
+  isProcessingPayment?: boolean
+  detectedPaymentMethod?: 'lightning' | 'onchain' | null
 }
 
 const feeRates = [
@@ -193,6 +49,8 @@ export const Step3: React.FC<StepProps> = ({
   paymentStatus,
   orderPayload,
   paymentReceived = false,
+  isProcessingPayment = false,
+  detectedPaymentMethod = null,
 }) => {
   const [paymentMethod, setPaymentMethod] = useState<'lightning' | 'onchain'>(
     'lightning'
@@ -504,12 +362,12 @@ export const Step3: React.FC<StepProps> = ({
         </div>
 
         {/* Payment Processing State */}
-        {localPaymentState === 'processing' && (
+        {(localPaymentState === 'processing' || isProcessingPayment) && (
           <OrderProcessingDisplay
             bitcoinUnit={bitcoinUnit}
             currentPayment={currentPayment}
             orderId={order?.order_id}
-            paymentMethod={paymentMethod}
+            paymentMethod={detectedPaymentMethod || paymentMethod}
           />
         )}
 
@@ -548,7 +406,8 @@ export const Step3: React.FC<StepProps> = ({
           localPaymentState !== 'success' &&
           paymentStatus !== 'success' &&
           localPaymentState !== 'error' &&
-          localPaymentState !== 'expired' && (
+          localPaymentState !== 'expired' &&
+          !isProcessingPayment && (
             <div className="bg-gray-900 text-white p-6 rounded-lg shadow-lg">
               <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
                 {/* Order Summary */}
@@ -598,7 +457,9 @@ export const Step3: React.FC<StepProps> = ({
                   )}
 
                   {/* QR Code Payment */}
-                  {!useWalletFunds && (
+                  {(!useWalletFunds ||
+                    (paymentMethod === 'lightning' &&
+                      outboundLiquidity <= 0)) && (
                     <div className="text-center">
                       {localPaymentState === 'waiting' ? (
                         <PaymentWaiting
@@ -631,7 +492,8 @@ export const Step3: React.FC<StepProps> = ({
         {/* Footer */}
         {localPaymentState !== 'processing' &&
           localPaymentState !== 'success' &&
-          paymentStatus !== 'success' && (
+          paymentStatus !== 'success' &&
+          !isProcessingPayment && (
             <div className="flex justify-center mt-6">
               <button
                 className="px-6 py-3 bg-gray-600 text-white rounded-lg hover:bg-gray-700 transition-colors font-medium"
