+++ src/routes/order-new-channel/components/index.ts
@@ -6,3 +6,5 @@ export { QRCodePayment } from './QRCodePayment'
 export { PaymentStatusDisplay } from './PaymentStatusDisplay'
 export { WalletPaymentSection } from './WalletPaymentSection'
 export { OrderProcessingDisplay } from './OrderProcessingDisplay'
+export { CountdownTimer } from './CountdownTimer'
+export { PaymentWaiting } from './PaymentWaiting'
